[Style] change code location
