+++ src/main/java/com/capstone/favicon/admin/application/service/StatisticsService.java
@@ -1,0 +1,11 @@
+package com.capstone.favicon.admin.application.service;
+
+import com.capstone.favicon.admin.dto.MonthlyCountDto;
+
+import java.util.List;
+import java.util.Map;
+
+public interface StatisticsService {
+    Map<String, Object> getUserCount();
+    List<MonthlyCountDto> getUserOverview();
+}
