+++ src/main/java/com/capstone/favicon/user/application/UserServiceImpl.java
@@ -5,7 +5,6 @@ import com.capstone.favicon.user.application.service.OTPService;
 import com.capstone.favicon.user.application.service.UserService;
 import com.capstone.favicon.user.domain.User;
 import com.capstone.favicon.user.dto.LoginDto;
-import com.capstone.favicon.user.dto.MonthlyCountDto;
 import com.capstone.favicon.user.dto.RegisterDto;
 import com.capstone.favicon.user.repository.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
@@ -16,11 +15,7 @@ import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.*;
-
-import static java.lang.Math.round;
-
 
 @RequiredArgsConstructor
 @Service
@@ -110,41 +105,6 @@ public class UserServiceImpl implements UserService {
             return true;
         }
         return false;
-    }
-
-    @Override
-    public Map<String, Object> getUserCount() {
-        LocalDateTime today = LocalDateTime.now();
-        LocalDateTime start = today.minusMonths(1).withDayOfMonth(1);
-        LocalDateTime end = start.plusMonths(1);
-
-        int previousCount = userRepository.countUsersAt(start, end);
-        int currentCount = userRepository.countUsersAt(end, today.plusDays(1));
-
-        double rate = 0.0;
-        if (previousCount > 0) {
-            rate = ((double) (currentCount-previousCount) / previousCount) *100.0;
-            rate = Math.round(rate*10.0)/10.0;
-        }
-
-        Map<String, Object> result = new HashMap<>();
-        result.put("total", userRepository.countAllUsers());
-        result.put("rate", rate);
-        return result;
-    }
-
-    @Override
-    public List<MonthlyCountDto> getUserOverview() {
-        LocalDateTime end = LocalDateTime.now().plusMonths(1).withDayOfMonth(1);
-        List<MonthlyCountDto> result = new ArrayList<>();
-        for (int i = 0; i<6; i++) {
-            LocalDateTime start = end.minusMonths(1);
-            int month = start.getMonthValue();
-            int count = userRepository.countUsersAt(start, end);
-            result.add(new MonthlyCountDto(month, count));
-            end = start;
-        }
-        return result;
     }
 
 }
