+++ src/main/java/com/capstone/favicon/user/application/service/UserService.java
@@ -1,12 +1,9 @@
 package com.capstone.favicon.user.application.service;
 
 import com.capstone.favicon.user.dto.LoginDto;
-import com.capstone.favicon.user.dto.MonthlyCountDto;
 import com.capstone.favicon.user.dto.RegisterDto;
 import jakarta.servlet.http.HttpServletRequest;
 
-import java.util.List;
-import java.util.Map;
 
 public interface UserService {
     void sendCode(RegisterDto.checkEmail checkEmail);
@@ -16,6 +13,4 @@ public interface UserService {
     void logout(HttpServletRequest request);
     void delete(HttpServletRequest request);
     boolean checkAdmin(HttpServletRequest request);
-    Map<String, Object> getUserCount();
-    List<MonthlyCountDto> getUserOverview();
 }
